requirements.txt
Django>=4.2,<5
djangorestframework
psycopg2-binary
django-cors-headers
python-dotenv
gunicorn
whitenoise
django-filter
Faker